requirements.txt
streamlit
openai
playwright
bs4
python-dotenv
